LayoutTests/editing/inserting/insert-paragraph-after-tab-span-and-text-expected.txt

This test ensures WebKit inserts a paragraph separator properly at the end of a tab span.
| "\n"
| <span>
|   class="Apple-tab-span"
|   style="white-space:pre"
|   "    hello"
| <div>
|   <#selection-caret>
|   <br>
|   "\n"
|   <span>
|     class="Apple-style-span"
|     <span>
|       class="Apple-tab-span"
|       style="white-space: pre; "
|       "    "
|     "world"
